use u16 for unit_mask. I'm not sure this catches *all* the instances, but it does appear to be an improvement -Graydon
Phil
rgards,

+++ dae/oprofiled.c
@@ -47,7 +47,7 @@ pid_t getpgid(pid_t pid);
 
 u32 ctr_count[OP_MAX_COUNTERS];
 u8 ctr_event[OP_MAX_COUNTERS];
-u8 ctr_um[OP_MAX_COUNTERS];
+u16 ctr_um[OP_MAX_COUNTERS];
 double cpu_speed;
 fd_t hashmapdevfd;
